got stuck on implementing current algorithm with the table in the docx file. Did some code refractoring before going to sleep.
